Add a GridBagLayout
Currently making good layouts in SWT applications is quite hard, even
though the BorderLayout already simplifies this but is most usefull for
layout cases with some fixed size widget are aranged around a central
widget.

This adds another Layout known from Swing that is the GridBagLayout wich
lets define one a number of bags organizeds in a grid where widgets take
relative sizes as well as margins and paddings, that way it is possible
to easily create layouts that react in a responsive way to size changes
of the container.